- fixed size_t/unsigned confusion
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@10104 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/components/ASSDriver/ASSDriver.cpp
@@ -365,7 +365,7 @@ ASSDriver::onUpdateSettings(dom::NodePtr theSettings) {
 y60::EventPtrList 
 ASSDriver::poll() {
     std::vector<unsigned char> myBuffer;
-    unsigned myByteCount = _mySerialPort->peek();
+    size_t myByteCount = _mySerialPort->peek();
     myBuffer.resize( myByteCount );
     _mySerialPort->read( reinterpret_cast<char*>(& ( * myBuffer.begin())), myByteCount );
